pyproject.toml
[project]
name = "market_sentiment"
version = "0.1.0"
description = "market_sentiment using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.4.1",
    "gnews>=0.2.1",
    "langchain-community>=0.0.17",
    "langchain-openai>=0.1.0",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "yfinance>=0.2.28",
    "python-dateutil>=2.8.2",
    "loguru>=0.7.2",
    "pandas>=2.1.3",
    "numpy>=2.1.3",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.2",
    "scikit-learn>=1.3.2",
    "scipy>=1.11.3",
    "statsmodels>=0.14.3",
    "python-dotenv>=1.0.0",
    "massive>=0.1.0",
]

[project.scripts]
market_sentiment = "market_sentiment.main:run"
run_crew = "market_sentiment.main:run"
train = "market_sentiment.main:train"
replay = "market_sentiment.main:replay"
test = "market_sentiment.main:test"
run_with_trigger = "market_sentiment.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

